fix runner.wait_for() to correctly behave on specified timeout
If timeout is given to observer.await_done(timeout=XXX)
it should be counted from that line of code
and not from observer.start_time

connection_observer.await_done() should not loop awaiting for not-None self._future
That is runner.wait_for() responsibility to take into account
that provided future may be None because of commands queue.
Runner should either await for future or timeout if future is not
created before final timeout of observer.